pyproject.toml
[project]
name = "future-agi-rag"
version = "0.1.0"
description = "Gemini RAG Document Q&A System with Future AGI Evaluation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.3.0",
    "traceai-google-genai>=0.1.1",
    "futureagi>=0.6.9",
    "ai-evaluation>=0.2.2",
    "chromadb>=0.4.22",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "tiktoken>=0.5.2",
    "gradio>=4.16.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "fi-instrumentation-otel>=0.1.16",
]
